Fixes encoder reading, taken from @jasper, + change encoder values to 32bit

+++ mirte_telemetrix_cpp/include/mirte_telemetrix_cpp/sensors/encoder_monitor.hpp
@@ -25,7 +25,7 @@ public:
   void data_callback(int16_t value);
 
 private:
-  std::atomic<int16_t> value = 0;
+  std::atomic<int32_t> value = 0;
 
   // Publisher: encoder/NAME
   rclcpp::Publisher<mirte_msgs::msg::Encoder>::SharedPtr encoder_pub;
